version 5.0

+++ controllers/user.controller.js
@@ -1,6 +1,7 @@
 
 const Users = require('../models/user-model');
-
+const path = require("path");
+const fs = require('fs');
 
 // get all Users
 const getAllUsers = async (req, res)=>{
@@ -41,7 +42,8 @@ const createUser = async (req, res)=>{
             name: name,
             email: email,
             phone: phone,
-            department: department
+            department: department,
+            image: req.file.filename,
         }
         const user = await Users.create(newUser);
         if(user){
@@ -61,14 +63,34 @@ const createUser = async (req, res)=>{
 const updateUser = async(req, res)=>{
     try {
         const {name, email, phone, department} = req.body;
-        let updateUser ={
+        const userId = req.params.id;
+
+        // find existing user
+        const existingUser = await Users.findById({_id: userId});
+        if (!existingUser) {
+            return res.status(404).send({ status: "fail", message: 'User not found' });
+        }
+
+        let updateNewUser ={
             name: name,
             email: email,
             phone: phone,
-            department: department
+            department: department,
         };
-        const userId = req.params.id
-        const user = await Users.findByIdAndUpdate({_id: userId}, {$set: updateUser}, {new:true});
+
+        // Delete the existing image if present
+        if (existingUser.image) {
+            const existingImagePath = path.join(__dirname, '../public/images', existingUser.image);
+            await fs.unlink(existingImagePath, (err) => {
+                if (err) console.error(`Failed to delete old image: ${err.message}`);
+            });
+        }
+
+        // Add new image to the update object
+        updateNewUser.image = req.file.filename;
+
+        // Update user data in the database
+        let user = await Users.findByIdAndUpdate({_id: userId}, {$set: updateNewUser}, {new:true});
         if(user){
             res.status(200).send({status: "success", message: 'user update successfully', data: user});
         }else{
@@ -84,6 +106,19 @@ const updateUser = async(req, res)=>{
 const deleteUser = async(req, res)=>{
     try {
         const userId = req.params.id;
+        let existingUser = await Users.findById({_id: userId});
+
+        if (!existingUser) {
+            return res.status(404).send({ status: "fail", message: 'User not found' });
+        }
+
+        // Delete the existing image if present
+        if (existingUser.image) {
+            const existingImagePath = path.join(__dirname, '../public/images', existingUser.image);
+            await fs.unlink(existingImagePath, (err) => {
+                if (err) console.error(`Failed to delete old image: ${err.message}`);
+            });
+        }
         const user = await Users.findByIdAndDelete({_id: userId});
         if(user){
             res.status(200).send({status: "success", message: 'user delete successfully', data: user});
